backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.5.0
Faker==18.11.2
filetype==1.2.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3